+++ src/Voucherify/DataModel/CustomerSummaryOrders.cs
@@ -8,13 +8,13 @@ namespace Voucherify.DataModel
     public class CustomerSummaryOrders
     {
         [JsonProperty(PropertyName = "total_amount")]
-        public int TotalAmount { get; private set; }
+        public long TotalAmount { get; private set; }
 
         [JsonProperty(PropertyName = "total_count")]
         public int TotalCount { get; private set; }
 
         [JsonProperty(PropertyName = "average_amount")]
-        public int? AverageAmount { get; private set; }
+        public long? AverageAmount { get; private set; }
 
         [JsonProperty(PropertyName = "last_order_amount")]
         public int? LastOrderAmount { get; private set; }
